Move auth related files into package
