+++ server/src/oes/registration/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Auth package."""
