+++ server/src/oes/registration/services/registration.py
@@ -3,7 +3,8 @@ from collections.abc import Iterable, Sequence
 from typing import Optional
 from uuid import UUID
 
-from oes.registration.entities.auth import AccountEntity
+from oes.registration.auth.entities import AccountEntity
+from oes.registration.auth.service import AuthService
 from oes.registration.entities.event_stats import EventStatsEntity
 from oes.registration.entities.registration import RegistrationEntity
 from oes.registration.hook.models import HookEvent
@@ -16,7 +17,6 @@ from oes.registration.models.registration import (
     SelfServiceRegistration,
 )
 from oes.registration.serialization import get_converter
-from oes.registration.services.auth import AuthService
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import contains_eager, selectinload
