+++ server/src/oes/registration/views/access_code.py
@@ -5,7 +5,7 @@ from typing import Optional
 from attrs import frozen
 from blacksheep import Response, auth
 from oes.registration.app import app
-from oes.registration.auth import RequireAdmin
+from oes.registration.auth.handlers import RequireAdmin
 from oes.registration.database import transaction
 from oes.registration.docs import docs, docs_helper
 from oes.registration.models.access_code import AccessCodeSettings
