+++ server/src/oes/registration/views/cart.py
@@ -21,12 +21,12 @@ from cattrs import BaseValidationError
 from oes.interview.response import IncompleteInterviewStateResponse
 from oes.interview.state import InvalidStateError
 from oes.registration.app import app
-from oes.registration.auth import RequireAdmin, RequireCart, RequireSelfService
+from oes.registration.auth.handlers import RequireAdmin, RequireCart, RequireSelfService
+from oes.registration.auth.models import User
 from oes.registration.database import transaction
 from oes.registration.docs import docs, docs_helper, serialize
 from oes.registration.entities.cart import CartEntity
 from oes.registration.entities.registration import RegistrationEntity
-from oes.registration.models.auth import User
 from oes.registration.models.cart import CartData, CartError, CartRegistration
 from oes.registration.models.config import Config
 from oes.registration.models.event import Event, EventConfig
