+++ server/src/oes/registration/views/event.py
@@ -4,9 +4,9 @@ from collections.abc import Sequence
 from blacksheep import auth
 from blacksheep.exceptions import NotFound
 from oes.registration.app import app
-from oes.registration.auth import RequireEvent
+from oes.registration.auth.handlers import RequireEvent
+from oes.registration.auth.models import User
 from oes.registration.docs import docs_helper
-from oes.registration.models.auth import User
 from oes.registration.models.event import Event, EventConfig
 from oes.registration.util import check_not_found
 from oes.registration.views.responses import EventResponse
